--- readme.md
@@ -1,10 +1,0 @@
-La aplicacion corre con el comando: npm run dev
-
-Teconolias.
-    1.- Express 
-    2.- Sequelize ORM
-    3.- Nodemon
-    4.- MySQL# CHECK-ROOM-API
-# CHECK-ROOM-API
-# CHECK-ROOM-APIREST
-# CHECK-ROOM-APIREST
